+++ HPXMLtoOpenStudio/measure.rb
@@ -293,6 +293,7 @@ class OSModel
 
     # Output
 
+    add_unmet_hours_output(runner, model, spaces)
     add_loads_output(runner, model, spaces, add_component_loads)
     set_output_files(runner, model)
     # Uncomment to debug EMS
@@ -1893,16 +1894,55 @@ class OSModel
     additionalProperties.setFeature('emissions_scenario_types', emissions_scenario_types)
   end
 
-  def self.map_to_string(map)
-    map_str = {}
-    map.each do |sys_id, objects|
-      object_name_list = []
-      objects.uniq.each do |object|
-        object_name_list << object.name.to_s
-      end
-      map_str[sys_id] = object_name_list if object_name_list.size > 0
+  def self.add_unmet_hours_output(runner, model, spaces)
+    # We do our own unmet hours calculation via EMS so that we can incorporate,
+    # e.g., heating/cooling seasons into the logic.
+    hvac_control = @hpxml.hvac_controls[0]
+    sim_year = @hpxml.header.sim_calendar_year
+
+    htg_start_day = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_begin_month, hvac_control.seasons_heating_begin_day)
+    htg_end_day = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_heating_end_month, hvac_control.seasons_heating_end_day)
+    clg_start_day = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_begin_month, hvac_control.seasons_cooling_begin_day)
+    clg_end_day = Schedule.get_day_num_from_month_day(sim_year, hvac_control.seasons_cooling_end_month, hvac_control.seasons_cooling_end_day)
+
+    living_zone = spaces[HPXML::LocationLivingSpace].thermalZone.get
+
+    # EMS sensors
+    htg_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Zone Heating Setpoint Not Met Time')
+    htg_sensor.setName('zone htg unmet s')
+    htg_sensor.setKeyName(living_zone.name.to_s)
+
+    clg_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Zone Cooling Setpoint Not Met Time')
+    clg_sensor.setName('zone clg unmet s')
+    clg_sensor.setKeyName(living_zone.name.to_s)
+
+    # EMS program
+    clg_hrs = 'clg_unmet_hours'
+    htg_hrs = 'htg_unmet_hours'
+    program = OpenStudio::Model::EnergyManagementSystemProgram.new(model)
+    program.setName(Constants.ObjectNameUnmetHoursProgram)
+    program.addLine("Set #{htg_hrs} = 0")
+    if htg_end_day > htg_start_day
+      program.addLine("If (DayOfYear >= #{htg_start_day}) && (DayOfYear <= #{htg_end_day})")
+    else
+      program.addLine("If (DayOfYear >= #{htg_start_day}) || (DayOfYear <= #{htg_end_day})")
     end
-    return map_str.to_s
+    program.addLine("  Set #{htg_hrs} = #{htg_hrs} + #{htg_sensor.name}")
+    program.addLine('EndIf')
+    program.addLine("Set #{clg_hrs} = 0")
+    if clg_end_day > clg_start_day
+      program.addLine("If (DayOfYear >= #{clg_start_day}) && (DayOfYear <= #{clg_end_day})")
+    else
+      program.addLine("If (DayOfYear >= #{clg_start_day}) || (DayOfYear <= #{clg_end_day})")
+    end
+    program.addLine("  Set #{clg_hrs} = #{clg_hrs} + #{clg_sensor.name}")
+    program.addLine('EndIf')
+
+    # EMS calling manager
+    program_calling_manager = OpenStudio::Model::EnergyManagementSystemProgramCallingManager.new(model)
+    program_calling_manager.setName("#{program.name} calling manager")
+    program_calling_manager.setCallingPoint('EndOfZoneTimestepBeforeZoneReporting')
+    program_calling_manager.addProgram(program)
   end
 
   def self.add_loads_output(runner, model, spaces, add_component_loads)